+++ src/schemas/CategorySchema.ts
@@ -6,7 +6,7 @@ export default class Category {
       id: "string",
       name: "string",
       color: "string",
-      number: "float",
+      number: "double",
     },
   };
 }
